+++ src/store/blog/getters.js
@@ -1,6 +1,6 @@
 import { COLLECTION_GETTERS } from '@/store/lib/mixins'
 
-// Post Module Getters
+// Blog Module Getters
 export default {
   ...COLLECTION_GETTERS
 }
