+++ src/store/blog/mutations.js
@@ -1,6 +1,6 @@
 import { COLLECTION_MUTATIONS } from '@/store/lib/mixins'
 
-// Post Module mutations
+// Blog Module mutations
 export default {
   ...COLLECTION_MUTATIONS
 }
